requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
fpdf==1.7.2
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
groq==0.11.0
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.8.4
numpy==2.1.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
plotly==5.24.1
proto-plus==1.24.0
protobuf==5.28.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
st-clickable-images==0.0.3
st-pages==1.0.1
streamlit==1.38.0
streamlit-card==1.0.2
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
